- Separate ds_task from ds_process. - Move ds_process back into worker. - Worker stores ds_process in separate table.
