+++ dataswarm/worker/Makefile
@@ -4,9 +4,9 @@ include ../../rules.mk
 TARGETS=$(PROGRAMS)
 PROGRAMS=ds_worker
 
-SOURCES=ds_worker_main.c ds_worker.c ds_task_table.c ds_blob_table.c
+SOURCES=ds_worker_main.c ds_worker.c ds_process.c ds_task_table.c ds_blob_table.c
 OBJECTS=$(SOURCES:%.c=%.o)
-DS_OBJECTS=../common/ds_message.o ../common/ds_process.o ../common/ds_resources.o ../common/ds_task.o ../common/ds_mount.o ../common/ds_blob.o
+DS_OBJECTS=../common/ds_message.o ../common/ds_resources.o ../common/ds_task.o ../common/ds_mount.o ../common/ds_blob.o
 
 LOCAL_CCFLAGS = -I ..
 LOCAL_LINKAGE = -L ../../dttools/src -ldttools
